+++ app/config/specs.py
@@ -1,17 +1,6 @@
-from app.build_specs.formatter_selectors import FormatterSelector, FormatterMapping
 from app.build_specs.mappings import BuildSpecMapping, PrepopBuildSpecMapping
 from app.build_specs.reader import BuildSpecFileRepository
 from app.build_specs.spec_selectors import BuildSpecSelector, BuildSpecPeriodSelector, PrepopSelector
-from app.formatters.cora_formatter import CORAFormatter, MESFormatter
-from app.formatters.cora_looping_formatter import CORALoopingFormatter
-from app.formatters.cs_formatter import CSFormatter
-from app.formatters.cs_looping_formatter import CSLoopingFormatter
-from app.formatters.formatter import Formatter
-from app.formatters.idbr_looping_formatter import IDBRLoopingFormatter
-from app.formatters.json_formatter import JSONFormatter
-from app.formatters.open_road_formatter import OpenRoadFormatter
-from app.formatters.spp_formatter import SPPFormatter
-from app.formatters.spp_looping_formatter import SPPLoopingFormatter
 
 
 build_spec_mapping: BuildSpecMapping = BuildSpecMapping({
@@ -57,14 +46,3 @@ prepop_spec_mapping: PrepopBuildSpecMapping = PrepopBuildSpecMapping({
     "221": PrepopSelector("bres-prepop"),
     "241": PrepopSelector("brs-prepop"),
 }, repository=BuildSpecFileRepository())
-
-
-formatter_mapping: FormatterMapping = FormatterMapping({
-    "CORA": FormatterSelector(CORAFormatter, CORALoopingFormatter),
-    "CORA_MES": FormatterSelector(MESFormatter),
-    "CS": FormatterSelector(CSFormatter, CSLoopingFormatter),
-    "OpenROAD": FormatterSelector(OpenRoadFormatter),
-    "SPP": FormatterSelector(SPPFormatter, SPPLoopingFormatter),
-    "IDBR": FormatterSelector(IDBRLoopingFormatter),
-    "JSON": FormatterSelector(JSONFormatter),
-})
