+++ app/routes.py
@@ -25,13 +25,6 @@ def process_pck(req: Request, _tx_id: TX_ID):
 
     logger.info("Received Parameters", **survey_metadata)
 
-    use_image_formatter: bool = req.args.get("use_image_formatter", "False").upper() == "TRUE"
-
-    if use_image_formatter:
-        logger.info("Using image transformer")
-    else:
-        logger.info("NOT using image transformer")
-
     for k, v in survey_metadata.items():
         if v == "":
             raise DataError(f"Missing required parameter {k} from request")
