+++ app/transform/execute.py
@@ -1,16 +1,9 @@
-from collections.abc import Callable
 from typing import Final
 
 from sdx_gcp.app import get_logger
 
+from app.config.functions import function_lookup
 from app.definitions import ParseTree, Transform, Field, Value, BuildSpecError
-from app.transform.functions.compound import currency_thousands, period_start, period_end
-from app.transform.functions.general import no_transform, exists, any_exists, lookup
-from app.transform.functions.lists import as_list, append_to_list, prepend_to_list, trim_list
-from app.transform.functions.numerical import round_half_up, aggregate, mean, number_equals, total, divide
-from app.transform.functions.string import starts_with, contains, any_contains, concat, carve, string_padding, \
-    space_split, postcode_start, postcode_end
-from app.transform.functions.time import to_date, any_date, start_of_month, end_of_month, start_of_year, end_of_year
 from app.transform.tree_walker import TreeWalker
 
 logger = get_logger()
@@ -18,42 +11,6 @@ logger = get_logger()
 
 DERIVED_PREFIX: Final[str] = "&"
 CURRENT_VALUE: Final[str] = DERIVED_PREFIX + "value"
-
-
-_function_lookup: dict[str, Callable] = {
-    "VALUE": no_transform,
-    "EXISTS": exists,
-    "ANY_EXISTS": any_exists,
-    "LOOKUP": lookup,
-    "STARTS_WITH": starts_with,
-    "CONTAINS": contains,
-    "ANY_CONTAINS": any_contains,
-    "CONCAT": concat,
-    "TO_DATE": to_date,
-    "ANY_DATE": any_date,
-    "START_OF_MONTH": start_of_month,
-    "END_OF_MONTH": end_of_month,
-    "START_OF_YEAR": start_of_year,
-    "END_OF_YEAR": end_of_year,
-    "ROUND": round_half_up,
-    "TOTAL": total,
-    "DIVIDE": divide,
-    "AGGREGATE": aggregate,
-    "MEAN": mean,
-    "NUMBER_EQUALS": number_equals,
-    "CURRENCY_THOUSANDS": currency_thousands,
-    "PERIOD_START": period_start,
-    "PERIOD_END": period_end,
-    "CARVE": carve,
-    "AS_LIST": as_list,
-    "APPEND_TO_LIST": append_to_list,
-    "PREPEND_TO_LIST": prepend_to_list,
-    "TRIM_LIST": trim_list,
-    "PADDING": string_padding,
-    "SPACE_SPLIT": space_split,
-    "POSTCODE_START": postcode_start,
-    "POSTCODE_END": postcode_end,
-}
 
 
 def execute(tree: ParseTree) -> dict[str, Value]:
@@ -84,7 +41,7 @@ def execute(tree: ParseTree) -> dict[str, Value]:
 
 def execute_transform(transform: Transform, walker: TreeWalker) -> Value | list[Value]:
     name = transform["name"]
-    f = _function_lookup.get(name)
+    f = function_lookup.get(name)
     if f is None:
         raise BuildSpecError(f"Transform name {name} is not a valid function!")
