+++ pyproject.toml
@@ -20,12 +20,9 @@ PyYAML = "6.0.1"
 
 
 [tool.poetry.group.dev.dependencies]
-flake8 = "7.1.0"
-
-[poetry.group.dev.dependencies]
+flake8 = "^7.1.0"
 pytest = "^7.0"
 pytest-cov = "^3.0"
-flake8 = "7.1.0"
 
 
 [build-system]
